tests/wpt/web-platform-tests/presentation-api/controlling-ua/getAvailability.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Getting the presentation displays availability information.</title>
<link rel="author" title="Marius Wessel" href="http://www.fokus.fraunhofer.de">
<link rel="author" title="Tomoyuki Shimizu" href="https://github.com/tomoyukilabs">
<link rel="help" href="http://w3c.github.io/presentation-api/#dfn-presentation-display-availability">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="common.js"></script>

<p>The test passes if a "PASS" result appears.</p>

<script>

    // ---------------------------------------
    // Presentation Availability Tests - begin
    // ---------------------------------------

    promise_test(function(t) {
        var availability;

        var request = new PresentationRequest(presentationUrls);
        assert_true(request instanceof PresentationRequest, 'The request is an instance of PresentationRequest.');

        var promise = request.getAvailability();
        assert_equals(promise, request.getAvailability(), 'If the PresentationRequest object has an unsettled Promise, getAvailability returns that Promise.');

        function catchNotSupported(err) {
            assert_equals(err.name, 'NotSupportedError', 'getAvailability() rejects a Promise with a NotSupportedError exception, if the browser can find presentation displays only when starting a connection.')
        }

        return promise.then(function (a) {
            availability = a;

            assert_true(availability instanceof PresentationAvailability, 'The promise is resolved with an instance of PresentationAvailability.');
            assert_equals(typeof availability.value, 'boolean', 'The availability has an boolean value.');
            assert_true(availability.value, 'The availability value is true when any presentation display is available.');

            var newPromise = request.getAvailability();
            assert_not_equals(promise, newPromise, 'If the Promise from a previous call to getAvailability has already been settled, getAvailability returns a new Promise.');

            return newPromise.then(function (newAvailability) {
                assert_equals(availability, newAvailability, 'Promises from a PresentationRequest\'s getAvailability are resolved with the same PresentationAvailability object.');
            }, catchNotSupported);
        }, catchNotSupported);
    });

</script>
